traits: Move {Split,Merge}Tag to `lofty::tag`
