+++ src/iff/wav/tag/mod.rs
@@ -3,8 +3,9 @@ mod write;
 
 use crate::config::WriteOptions;
 use crate::error::{LoftyError, Result};
-use crate::tag::{try_parse_year, Accessor, ItemKey, ItemValue, Tag, TagExt, TagItem, TagType};
-use crate::traits::{MergeTag, SplitTag};
+use crate::tag::{
+	try_parse_year, Accessor, ItemKey, ItemValue, MergeTag, SplitTag, Tag, TagExt, TagItem, TagType,
+};
 
 use std::borrow::Cow;
 use std::fs::File;
